Refactor UI into new package
